Add Vulkan Renderer backend (#22)
* Add VulkanRendererAPI

* Add Vulkan to project

* Remove Cabrankeditor project

* Add Vulkan as API in enum

* Update formatting

* Add Vulkan classes

* Add Vulkan shaders

* Add VUlkan sdk to project

* Refactor OpenGL to work alongside Vulkan

* Refactor Material pipeline

* Adapt engine for Vulkan

* Fixed Asset converter for multi mesh models

* Fix slang shaders (WIP)

* Finished PBR.slang

* Polish Application/Window/Layers ownerships

* Fix logging macros, SandboxApp layer creation and MaterialKind enum

* Move Math library

* Fix RendererAPI leak

* Polish high-level Renderer

* Clean inclusions and fix material fallback

* Fix inclusions and guards

* Polish Vulkan result checking

* Vulkan polish

* Add options to build different renderers

* Fix batch rendering

* Fix OpenGL rendering

* Add comments to shaders

* Add Storage Buffer for Vulkan and fix fixed point light amount

* Polish Asset converter

* Fix slang shader

* Add mipmams functionallity

* Update documentation

+++ Cabrankengine/src/Platform/OpenGL/OpenGLTexture.h
@@ -23,7 +23,7 @@ namespace cbk::platform::opengl {
 			virtual uint32_t getHeight() const override { return m_Height; }
 		
 			// Returns the texture's renderer ID, which is used by the graphics API to identify the texture.
-			virtual uint32_t getRendererID() const override { return m_RendererID; }
+			virtual uint64_t getRendererID() const override { return m_RendererID; }
 
 			// Returns the path to the texture file.
 			virtual const std::string& getPath() const override { return m_Path; }
